Rewrote my old code, and added a disconnect button.
Most (except resizing) old functionality is back, the simple settings page is gone because I quickly and sloppily implemented it - when it's not really needed, and there is now a disconnect button.

There is a bug where I can't put an svg in a certain button? I have absolutely no idea why.